+++ templates/admin/dashboard.php
@@ -7,7 +7,7 @@
 	<form action="admin-post.php" method="post">
 		<h3><?php _e('Jigoshop Dashboard', 'jigoshop'); ?></h3>
 
-		<p id="wp-version-message"><?php _e('You are using', 'jigoshop'); ?>
+		<p id="wp-version-message"><?php _e('You are using version:	', 'jigoshop'); ?>
 			<strong><?php echo \Jigoshop\Core::VERSION; ?></strong>
 		</p>