+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 * 2.0.1:
     * Fix: Migration alert message.
+    * Fix: Updated Jigoshop logotypes.
+    * Fix: Admin Order and Product list search.
 * 2.0.0 - 2016.10.27:
     * Full rewrite of core.